refactoring, - merge api.routes.js,scaling.js into app.js,main.js (which will be cleaned up / reduced later on)
